requirements.txt
Flask==2.3.1
ffmpeg==1.4
pydub==0.25.1
google-cloud-speech
google-api-python-client
google-cloud-storage
google-cloud-translate
google-cloud-texttospeech